server/requirements.txt
flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
PyGithub==2.1.1
shortuuid==1.0.11
